Move middlewares from app to middleware module while I'm at it
