+++ server/src/middleware.ts
@@ -1,6 +1,7 @@
 import { Response, Request, NextFunction } from "express";
 import bodyParser from "body-parser";
 import { getApiKeys, getApiSunset } from "./config";
+import { logger } from "./logger";
 import { absoluteUrl } from "./utils";
 
 export interface AppRequest extends Request {
@@ -87,5 +88,38 @@ export function addSunsetHeaders(
     response.header("Sunset", sunset.date.toHTTP());
     response.header("Link", `<${infoUrl}>;rel="sunset";type="text/html"`);
   }
+  next();
+}
+
+/**
+ * Set the Cache-Control HTTP response header's max-age directive to the given
+ * number of seconds.
+ */
+export function cacheControlMaxAge(seconds: number) {
+  return function (
+    request: AppRequest,
+    response: Response,
+    next: NextFunction
+  ): void {
+    if (request.method == "GET") {
+      const directives = [
+        request.authorization ? "private" : "public",
+        `max-age=${seconds}`,
+      ];
+      response.set("Cache-Control", directives.join(", "));
+    }
+    next();
+  };
+}
+
+/**
+ * Log basic info about the request at DEBUG level.
+ */
+export function logRequest(
+  request: Request,
+  response: Response,
+  next: NextFunction
+): void {
+  logger.debug(`${response.statusCode} ${request.method} ${request.url}`);
   next();
 }
